Refactoring and modularization
