+++ src/util/get-scopes.js
@@ -3,18 +3,17 @@
 const vscode = require('vscode');
 const hscopes = vscode.extensions.getExtension('yfzhao.hscopes-booster');
 
-/**
- * get TextMate Scope
- * @param {vscode.TextDocument} document 
- * @param {vscode.Position} position 
- * @returns { string[] | undefined}
- */
-function getScope(document, position) {
-    if (!hscopes || !document || !position){
-        console.log(`function "getScope" causes error.`)
-        return undefined
+// Caches the most recent scoping results
+let lastScopeCache = null;
+function getCachedMathScope(document, position) {
+    const key = `${document.uri.toString()}:${position.line}:${position.character}`;
+    if (!lastScopeCache || lastScopeCache.key !== key) {
+        lastScopeCache = {
+            key,
+            value: getMathScope(document, position)
+        };
     }
-    return  hscopes?.exports?.getScopeAt(document, position);
+    return lastScopeCache.value;
 }
 
 /**
@@ -23,26 +22,26 @@ function getScope(document, position) {
  * @param {vscode.Position} position 
  */
 function getMathScope(document, position) {
-    if (!document.languageId.match(/latex|markdown/)) 
+    if (!document.languageId.match(/latex|markdown/))
         return undefined
 
     let isInBeginDelimiter = false;
     let isInEndDelimiter = false;
-    let isDisplayMath = false; 
+    let isDisplayMath = false;
     // get scope 
     const scopeInfo = getScope(document, position)
-    const scope =  scopeInfo?.scopes?.toString();
+    const scope = scopeInfo?.scopes?.toString();
     const scopeRange = scopeInfo?.range;
 
     // not in math environment || can't get 'dravin.hscopes' extension
     // get isDisplayMath
-    if (scope?.includes("support.class.math")){
+    if (scope?.includes("support.class.math")) {
         isDisplayMath = scope.includes("math.block.environment")
     }
-    else if(scope?.includes("markup.math") || scope?.includes("markup.inserted.math")){
+    else if (scope?.includes("markup.math") || scope?.includes("markup.inserted.math")) {
         isDisplayMath = scope.includes("math.block") || scope.includes("math.display")
     }
-    else{
+    else {
         return undefined
     }
 
@@ -50,28 +49,27 @@ function getMathScope(document, position) {
     if (scope.includes("definition.string.begin")) {          // inBeginDelimiter
         isInBeginDelimiter = true
     }
-    else if (scope.includes("definition.string.end") ) {       // inEndDelimiter
+    else if (scope.includes("definition.string.end")) {       // inEndDelimiter
         isInEndDelimiter = !position.isEqual(scopeRange.start)
     }
-    else if (document.languageId === 'markdown' && scope.includes("definition.math") ) { // in delimiters!!!
+    else if (document.languageId === 'markdown' && scope.includes("definition.math")) { // in delimiters!!!
         if (position.isEqual(scopeRange.start)) {
             // (0,0) => true
-            if (!position.character && !position.line) 
+            if (!position.character && !position.line)
                 isInBeginDelimiter = true;
             // test previous character
             else {
-                const previousPosition = !!position.character 
-                                    ? new vscode.Position(position.line,position.character-1) 
-                                    : document.lineAt(position.line-1).range.end;
-                const previousPositionScope = getScope(document,previousPosition).scopes.toString()
-                isInBeginDelimiter = !(previousPositionScope.includes("math") || scope.includes("definition.math.end") ) 
-                                    || previousPositionScope.includes("definition.math.end") 
+                const previousPosition = !!position.character
+                    ? new vscode.Position(position.line, position.character - 1)
+                    : document.lineAt(position.line - 1).range.end;
+                const previousPositionScope = getScope(document, previousPosition).scopes.toString()
+                isInBeginDelimiter = !(previousPositionScope.includes("math") || scope.includes("definition.math.end"))
+                    || previousPositionScope.includes("definition.math.end")
             }
         }
         else {
-            // TODO: identify begin-delimiter and end delimiter
             isInEndDelimiter = true  // BUG: mix begin-delimiter and end-delimiter
-            isInBeginDelimiter = !scope.includes("definition.math.end") 
+            isInBeginDelimiter = !scope.includes("definition.math.end")
         }
     }
     return {
@@ -82,8 +80,28 @@ function getMathScope(document, position) {
     }
 }
 
+/**
+ * get TextMate Scope
+ * @param {vscode.TextDocument} document 
+ * @param {vscode.Position} position 
+ * @returns { string[] | undefined}
+ */
+function getScope(document, position) {
+    if (!hscopes || !document || !position) {
+        console.log(`function "getScope" causes error.`)
+        return undefined
+    }
+    return hscopes?.exports?.getScopeAt(document, position);
+}
+
+function isValidMathScope(scope) {
+    return scope && !scope.isInBeginDelimiter && !scope.isInEndDelimiter;
+}
+
 /// TODO: return match delimiters
 
 module.exports = {
-    getMathScope
+    getCachedMathScope,
+    getMathScope,
+    isValidMathScope
 }
