+++ src/util/texRenderer.js
